--- Staubsauger_JavaEdition/src/Field.java
@@ -1,12 +1,0 @@
-public class Field {
-    public int[] coords = new int[2];
-    public boolean blocked = false;
-    public int index;
-
-    Field(int pX, int pY, boolean pBlocked, int pIndex) {
-        coords[0] = pX;
-        coords[1] = pY;
-        blocked = pBlocked;
-        index = pIndex;
-    }
-}
