+++ Staubsauger_JavaEdition/src/world/Coordinate.java
@@ -1,0 +1,10 @@
+package world;
+
+public class Coordinate {
+    public int x, y;
+
+    public Coordinate(int pX, int pY) {
+        x = pX;
+        y = pY;
+    }
+}
