+++ Staubsauger_JavaEdition/src/world/Field.java
@@ -1,0 +1,13 @@
+package world;
+
+public class Field {
+    public Coordinate coord;
+    public boolean blocked = false;
+    public int index;
+
+    Field(int x, int y, boolean pBlocked, int pIndex) {
+        coord = new Coordinate(x, y);
+        blocked = pBlocked;
+        index = pIndex;
+    }
+}
